Changed money data type to int as we only use whole numbers Fixed the upgrade system Fixed syntax to ensure consistency Fixed the inconsistency with upgrade not consuming a turn - now upgrading a property uses your turn (can try to upgrade and if false you still roll), viewing portfolio no longer uses your turn Added method to search by field ID specifically (for upgrade system).

+++ Testing/game/files/PlayerTesting.java
@@ -11,8 +11,8 @@ class PlayerTesting {
 	int invalidPlayerIdLower, invalidPlayerIdUpper;
 	String validPlayerNameLowerBoundary, validPlayerNameUpperBoundary, validPlayerNameLower, validPlayerNameUpper;
 	String invalidPlayerNameLower, invalidPlayerNameUpper;
-	double validMoneyBoundary, validMoney;
-	double invalidMoney;
+	int validMoneyBoundary, validMoney;
+	int invalidMoney;
 	int validBoardPositionLowerBoundary, validBoardPositionUpperBoundary, validBoardPositionLower,
 			validBoardPositionUpper;
 	int invalidBoardPositionLower, invalidBoardPositionUpper;
